refactor: hygiene cleanup (completion comment, Screen enum split, effWPM note)
Minor cleanup: clarify completion check comment, split Screen enum into dedicated
file, add note on effective WPM formula in new-best calculation, update roadmap.
